+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Best effort list of major changes and bugfixes
 - feat: `local` source replaces `file`.
 - bugfix: standardize `resource-only` behaviour to always host-reload
 - feat: enable shell completion
+- feat: OCI sources now support rollback.
 
 ## 0.7.1
 - bugfix: services are only enabled/disabled when static